requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.7
dnspython==2.3.0
Flask==2.2.3
Flask-SocketIO==5.3.3
frozenlist==1.3.3
gptcache==0.1.14
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.142
loguru==0.7.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.1
pydantic==1.10.7
python-dateutil==2.8.2
python-engineio==4.4.0
python-socketio==5.8.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.47
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
win32-setctime==1.1.0
yarl==1.8.2
gevent-websocket
tiktoken==0.3.3


